+++ HISTORY.md
@@ -78,6 +78,9 @@
     * German
     * French
 
+### 2019-05-15
+* Objects, Treads: refactored videoLayer so it can also be used for other extensions (maps, 3d)
+
 ### 2019-05-14
 * Objects: fixed originalCloneName reference when setting a new parent